ISPN-3226 Enable various reports so that CI can show them
* PMD
* CPD
* SpotBugs (replaces FindBugs)
* Dependency Check
* Enable reports for branch builds

+++ commons/src/main/java/org/infinispan/commons/util/ProcessorInfo.java
@@ -34,7 +34,7 @@ public class ProcessorInfo {
    public static int availableProcessors() {
       if (System.getSecurityManager() != null) {
          return AccessController.doPrivileged((PrivilegedAction<Integer>) () ->
-               Integer.valueOf(org.infinispan.commons.jdkspecific.ProcessorInfo.availableProcessors()).intValue());
+               org.infinispan.commons.jdkspecific.ProcessorInfo.availableProcessors());
       }
 
       return org.infinispan.commons.jdkspecific.ProcessorInfo.availableProcessors();
